chore: moving peerset view to its own file and introduce `getTarget` method
