+++ dot/network/request_response.go
@@ -43,7 +43,7 @@ func (rrp *RequestResponseProtocol) Do(to peer.ID, req Message, res ResponseMess
 
 	defer func() {
 		err := stream.Close()
-		if err != nil && err.Error() != "stream reset" {
+		if err != nil && err.Error() != ErrStreamReset.Error() {
 			logger.Warnf("failed to close stream: %s", err)
 		}
 	}()
